Move Composite into own file
